+++ docs/source/changelog.rst
@@ -6,8 +6,9 @@ Version 1.NEXT
 
 *Release date: tbd*
 
+* added: **multithreading** for much faster torrent creation (about 30-40% faster). The number of threads can be controlled with the new ``--threads`` option. It defaults to using 4 threads which will be a good choice on most systems.
 * added: ``--version`` command
-* improved formatting and content of ``--help`` output
+* improved: formatting and content of ``--help`` output
 
 Version 1.0.1
 -------------